dropped obj files
