+++ src/WMISupport.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.1.31903.286
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WMISupport", "WMISupport.csproj", "{431B3E04-932A-4BD0-810A-7EC3B1C23F5A}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{431B3E04-932A-4BD0-810A-7EC3B1C23F5A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{431B3E04-932A-4BD0-810A-7EC3B1C23F5A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{431B3E04-932A-4BD0-810A-7EC3B1C23F5A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{431B3E04-932A-4BD0-810A-7EC3B1C23F5A}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {887B6D3F-905E-456D-8023-1ADF7E163DF0}
+	EndGlobalSection
+EndGlobal
